pyproject.toml
[tool.poetry]
name = "corsar-wallet"
version = "0.1.0"
description = ""
authors = ["Ivan Piankov <vanypyankov@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.95.0"
alembic = "^1.9.2"
environs = "^9.5.0"
inject = "^4.3.1"
uvicorn = "^0.20.0"
orjson = "^3.8.5"
sqlalchemy = "^2.0.0"
sqlalchemy2-stubs = "^0.0.2a32"
fastapi-jwt-auth = "^0.5.0"
passlib = "^1.7.4"
jose = "^1.0.0"
mashumaro = "^3.5"
psycopg2-binary = "^2.9.5"
pydantic = {extras = ["email"], version = "^1.10.6"}
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
asyncpg = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
pytest = "*"
isort = "*"
autopep8 = "^2.0.2"


[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git
    | \.env
    | \.venv
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
